changing from unsigned to site_t

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -7,7 +7,7 @@
 */
 size_t print_list(const list_t *h)
 {
-	unsigned int con;
+	size_t con;
 	const list_t *ptr;
 
 	ptr = h;
